API/pyproject.toml
[tool.poetry]
name = "moneyplant"
version = "0.1.0"
description = "Bin Controller Software"
authors = ["Christo <christotantony2003@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.75.1"
uvicorn = {extras = ["standard"], version = "^0.17.6"}
SQLAlchemy = "^1.4.35"
mariadb = "^1.0.11"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-multipart = "^0.0.5"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
